requirements.txt
opentelemetry-api>=1.30,<2.0
opentelemetry-exporter-otlp-proto-http>=1.30,<2.0
opentelemetry-sdk>=1.30,<2.0
fastapi>=0.115,<1.0
langchain>=0.3,<0.4
langchain-openai>=0.3,<0.4
pydantic>=2.7,<3.0
requests>=2.32,<3.0
uvicorn[standard]>=0.34,<1.0
